refactor: Restructure repository to standard Python src/ layout
STEP 1: Core Structure Creation
- Created src/, src/sincor2/, tests/, docs/, examples/ directories

STEP 2: Code Migration and Package Initialization
- Moved all 31 Python source files from root to src/sincor2/ using git mv
- Created src/sincor2/__init__.py with package metadata and exports

STEP 3: Dependency and Configuration Update
- Created modern pyproject.toml with setuptools build system
- Includes all dependencies from requirements.txt
- Added optional extras (dev, redis, docs)

STEP 4: Import Refactoring (CRITICAL)
- Refactored 7 intra-package imports in src/sincor2/
- Refactored 7 imports in tests/test_value.py
- Updated sys.path logic for test discovery from src/

STEP 5: Clean Up
- Created placeholder __init__.py for tests/
- Created README.md for docs/ and examples/

All changes tracked with git mv to preserve history.
