+++ docs/README.md
@@ -1,0 +1,5 @@
+"""
+SINCOR2 Documentation
+
+This directory contains documentation for the sincor2 platform.
+"""
