+++ examples/README.md
@@ -1,0 +1,10 @@
+# Examples
+
+This directory contains example scripts and notebooks demonstrating how to use the SINCOR2 platform.
+
+## Contents
+
+- Example configurations
+- Integration examples  
+- Demo scripts
+- Jupyter notebooks (if any)
