+++ pyproject.toml
@@ -1,0 +1,114 @@
+[build-system]
+requires = ["setuptools>=68.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "sincor2"
+version = "1.0.0"
+description = "SINCOR Master Platform - A comprehensive platform for intelligent commerce, monetization, and content management."
+readme = "README.md"
+requires-python = ">=3.9"
+license = {text = "MIT"}
+authors = [
+    {name = "OrderofChaos33", email = ""}
+]
+keywords = ["commerce", "monetization", "ai", "flask", "ecommerce"]
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "Intended Audience :: Developers",
+    "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
+]
+
+dependencies = [
+    "Flask>=3.0.0,<4.0.0",
+    "requests>=2.31.0",
+    "PyJWT>=2.8.0",
+    "Jinja2>=3.1.0",
+    "PyYAML>=6.0",
+    "python-dotenv>=1.0.0",
+    "numpy>=1.24.0",
+    "flask-jwt-extended>=4.7.0",
+    "pydantic[email]>=2.0.0",
+    "email-validator>=2.0.0",
+    "flask-limiter>=3.8.0",
+    "anthropic>=0.69.0",
+    "gunicorn>=21.0.0",
+    "psutil>=5.9.0",
+    "Werkzeug>=3.0.0",
+]
+
+[project.optional-dependencies]
+dev = [
+    "pytest>=7.0.0",
+    "pytest-cov>=4.0.0",
+    "black>=23.0.0",
+    "isort>=5.12.0",
+    "flake8>=6.0.0",
+    "mypy>=1.0.0",
+]
+redis = [
+    "redis>=5.0.0",
+]
+docs = [
+    "sphinx>=6.0.0",
+    "sphinx-rtd-theme>=1.2.0",
+]
+
+[project.urls]
+Homepage = "https://github.com/OrderofChaos33/sincor2"
+Repository = "https://github.com/OrderofChaos33/sincor2.git"
+Documentation = "https://github.com/OrderofChaos33/sincor2"
+Issues = "https://github.com/OrderofChaos33/sincor2/issues"
+
+[project.scripts]
+sincor2 = "sincor2.app:main"
+
+[tool.setuptools]
+packages = ["sincor2"]
+package-dir = {"" = "src"}
+
+[tool.black]
+line-length = 100
+target-version = ['py39', 'py310', 'py311', 'py312', 'py313']
+include = '\.pyi?$'
+extend-exclude = '''
+/(
+  # directories
+  \.eggs
+  | \.git
+  | \.hg
+  | \.mypy_cache
+  | \.tox
+  | \.venv
+  | build
+  | dist
+)/
+'''
+
+[tool.isort]
+profile = "black"
+line_length = 100
+multi_line_mode = 3
+include_trailing_comma = true
+force_grid_wrap = 0
+use_parentheses = true
+ensure_newline_before_comments = true
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
+python_files = ["test_*.py", "*_test.py"]
+addopts = "-v --tb=short"
+
+[tool.mypy]
+python_version = "3.9"
+warn_return_any = true
+warn_unused_configs = true
+disallow_untyped_defs = false
+disallow_incomplete_defs = false
+check_untyped_defs = false
