+++ src/sincor2/app.py
