+++ src/sincor2/dynamic_pricing_engine.py
