+++ src/sincor2/infinite_scaling_engine.py
