+++ src/sincor2/partnership_framework.py
