+++ src/sincor2/persona_engine.py
