+++ src/sincor2/recursive_value_products.py
