+++ src/sincor2/security_headers.py
