+++ src/sincor2/swarm_coordination.py
