+++ tests/__init__.py
@@ -1,0 +1,5 @@
+"""
+SINCOR2 Test Suite
+
+This package contains all test modules for the sincor2 platform.
+"""
