+++ tests/test_sincor.py
